Move mock objects into a dedicated module
