+++ tests/test_modern.py
@@ -11,9 +11,10 @@ try:
     from pathlib import Path
     from unittest.mock import Mock, call, patch
 except ImportError:  # Python 2.7, PyPy2
-    from mock import patch
+    pytest.importorskip("pathlib")
 
 from cli_test_helpers import ArgvContext
+from py3_mocks import DirectoryMock, FileMock, SymlinkMock
 
 import pyclean.cli
 import pyclean.modern
@@ -23,56 +24,6 @@ from pyclean.modern import (
     remove_debris_for,
     remove_freeform_targets,
 )
-
-
-class FilesystemObjectMock(Mock):
-
-    def __init__(self, *args, **kwargs):
-        name = kwargs['name'] if 'name' in kwargs else args[0]
-        super().__init__(autospec=Path(name), **kwargs)
-        self.name = name
-
-    def __eq__(self, other):
-        return self.name == other.name
-
-    def __lt__(self, other):
-        return self.name < other.name
-
-    def is_file(self):
-        return False
-
-    def is_dir(self):
-        return False
-
-    def is_symlink(self):
-        return False
-
-
-class DirectoryMock(FilesystemObjectMock):
-
-    def __init__(self, **kwargs):
-        super().__init__('a-dir', **kwargs)
-
-    def is_dir(self):
-        return True
-
-
-class FileMock(FilesystemObjectMock):
-
-    def __init__(self, **kwargs):
-        super().__init__('a-file', **kwargs)
-
-    def is_file(self):
-        return True
-
-
-class SymlinkMock(FilesystemObjectMock):
-
-    def __init__(self, **kwargs):
-        super().__init__('a-symlink', **kwargs)
-
-    def is_symlink(self):
-        return True
 
 
 @pytest.mark.skipif(sys.version_info < (3,), reason="requires Python 3")
